move helper class to utils module
